fingerprint/requirements.txt
flask
celery[redis]
ffmpeg-python
imagehash
Pillow
librosa
numpy
elasticsearch